s4:torture:smb2: avoid leaking tree connects up to the main function from the durable_open test
The tree connects are handed in from the calling wrapper.
Those that are freed and reconnected inside the test function
can not be freed in the wrapper and stick to the torture_context
until this is released in the main function.

Autobuild-User: Michael Adam <obnox@samba.org>
Autobuild-Date: Fri Oct 28 02:37:25 CEST 2011 on sn-devel-104