s4:torture:smb2: fix the durable_open test to succeed against w7 and w2k8r2
When a first client that has a durable open with share read/write/delete
and a read-write-handle lease on the file disconnects, a second
client will succeed in opening the file and the new client will be
given a RWH-lease if requested, not only a RH-lease, as was previously
checked in the test.

This might have been a bug in win7 build 7000, which is what the
comments in the test give as reference.