s3-nmbd Remove AD netlogon response from s3 nmbd server
I do not want users with misconfigurations to have nmbd respond
with this partially correct packet.  For example, it hardcodes
the site as Default-First-Site-Name.

If nmbd wishes to return this information, it would need to query
the AD database using the same APIs that the source4/ nbt server
does.

Andrew Bartlett

Autobuild-User: Andrew Bartlett <abartlet@samba.org>
Autobuild-Date: Thu Nov 17 02:10:54 CET 2011 on sn-devel-104