s3:vfs:gpfs remove non-working code
when development of vfs_gpfs was started a long time ago it was
considered a good idea to have it compileable even if GPFS is not
around and then the module would complain in the logs that libgpfs
is not there

Nowadays this does not work any more as the code contains GPFS
specific types that need the GPFS header files.

Instead of making it compile again by adding GPFS typedefs, this
removes the whole approach. Other VFS modules also do not even
compile when necessary headers are not around, so do the same for
the GPFS module code.