Add DRSUAPI_DRS_UPDATE_(x) flags
DRSUAPI_DRS_UPDATE flags are used in
DRS_MSG_REPMOD_V1 message structure when repsFrom
is modified via RPC.  The RPCs are currently uncoded but
samba_kcc maintains the flags (and uses them to identify
what repsFrom changes are to be executed).  These are currently
helpful to samba_kcc and are intended to ultimately be used in
RPCs.

Signed-off-by: Andrew Tridgell <tridge@samba.org>