s3-winbind: Add an update function for winbind cache.
With 57b3d32 we changed the format for the winbind cache database and
the code deleted the database for the upgrade. As this database holds
also cached credentials, removing it is not an option. We need to update
from version 1 to version 2.

Fix bug #8643 (A Samba update to 3.6 deletes the winbind cache with cached
credentials).
(cherry picked from commit df616e928e08646bbc234af673a496117535c6d8)