s3-winbindd Only use SamLogonEx when we can get unencrypted session keys
This ensures that we have some check on the session keys being returned
as the RC4 cipher is not checksumed.

The check comes from the fact that the credentials chain is tied to
the netlgon session key, and so if the credentials check passes then
the netlogon session key will be correct, and so the user session key
will be correctly decrypted.

Andrew Bartlett

Signed-off-by: Matthieu Patou <mat@matws.net>

s3: If we can't do validation 6 or sam_logon_ex use sam_logon only
(cherry picked from commit c119cd8868fc7e2eb08b09f7092519007fd83bf6)