selftest: Fix the exit code of make test when run in buildfarm
When make test is run in buildfarm environment (RUN_FROM_BUILD_FARM=1),
last process in make test pipeline is subunit-filter. This subunit-filter
is responsible for converting failure test result to xfail for known
failures. While converting failures to xfail, it sets exitcode=1. If
this is the last process in the pipeline make test fails, which is
the observed behavior in buildfarm.

In autobuild, there is additional process in the pipeline format-subunit
which processes the output of subunit-filter and correctly terminates
with exit code=0. The buildfarm make test failure is fixed by adding
extra filter-subunit in the pipeline, as the extra filter-subunit
exits correctly with exit code=0.

Autobuild-User: Amitay Isaacs <amitay@samba.org>
Autobuild-Date: Fri Feb  3 07:00:04 CET 2012 on sn-devel-104