smbwrapper: Remove smbwrapper
There are now many better ways to access a remote SMB filesystem,
which do not rely on LD_PRELOAD and the associated dangers.  FUSE,
gvfs and the CIFS VFS are all much better options which do not require
knowing every possible libc entry point that can deal with a file
descriptor.

As an example of the maintainence that would be required to keep this
going, recent changes to deal with thread races and close-on-exec have
resulted in dup3(), but this isn't currently mapped.  While this would
not be hard to add, it illistrates why it is better to move to an
interface designed for this task.

Andrew Bartlett

Autobuild-User: Andrew Bartlett <abartlet@samba.org>
Autobuild-Date: Thu Feb  9 01:58:24 CET 2012 on sn-devel-104