heimdal_build: omit #line statments
This restores and finishes my original commit
80e23c68d83a7c9989f87d5a88a78bb76d222afc, reverted in
68c61a829b8487104483b23052b54c532fecb6ce

    heimdal_build omit #line statments to allow valgrind to work again

This time however, the reason to omit line statements is that it
causes more trouble with the lcov code coverage system than the (nil)
value that these statements bring.  Otherwise, we have to have a
special case to remove the .gcno and .gcda files for these generated
files.

Andrew Bartlett