s4:torture:smb2: add a durable-open.open-oplock test
this is a variant of the open-lease test that does the initial
durable open with a batch oplock instead of a RH lease.

This is e.g. useful do analyse a (mis?) behaviour with current
windows 8 preview versions, compared to w2k8r2

Autobuild-User: Michael Adam <obnox@samba.org>
Autobuild-Date: Tue Feb 21 19:51:21 CET 2012 on sn-devel-104