selftest: skip the troublesome samba4.rpc.unixinfo test
The issue here is that while the single rpc_server process is stuck in
an nss_winbind getpwuid() call, winbindd cannot contact netlogon to
make the connection to the domain.

nss_winbind comes into play when (for s3fs) the
NSS_WRAPPER_WINBIND_SO_PATH environment variable is set.  In the
medium term, the unixinfo pipe should either be rewritten fully async
or removed.

Andrew Bartlett